feat(weight): added date filter, extracted weight chart into component
