+++ app/src/routes/(connected)/weight/+page.svelte
@@ -1,209 +1,22 @@
 <script lang="ts">
 	import { enhance } from '$app/forms';
 	import WeightListDisplay from '$lib/components/custom/weight/WeightListDisplay.svelte';
+	import WeightChart from './components/WeightChart.svelte';
 	import { Button } from '$lib/components/ui/button/index.js';
 	import * as Dialog from '$lib/components/ui/dialog/index.js';
 	import * as InputGroup from '$lib/components/ui/input-group/index.js';
 	import type { Weight } from '$lib/server/db/schema';
-	import {
-		createDeferred,
-		dateToStringChartTimeScaleFormatted,
-		dateToStringCustomFormat,
-		enhanceWithParam
-	} from '$lib/util.js';
-	import {
-		BarController,
-		BarElement,
-		CategoryScale,
-		Chart,
-		Legend,
-		LinearScale,
-		LineController,
-		LineElement,
-		PointElement,
-		TimeScale,
-		Title,
-		Tooltip
-	} from 'chart.js';
-	import 'chartjs-adapter-date-fns';
-	import { onMount } from 'svelte';
+	import { createDeferred, dateToStringCustomFormat, enhanceWithParam } from '$lib/util.js';
 	import { toast } from 'svelte-sonner';
-	import { SvelteMap } from 'svelte/reactivity';
 	import type { ActionData, PageServerData } from './$types.js';
-
-	Chart.register(
-		LineElement,
-		PointElement,
-		LineController,
-		BarController,
-		BarElement,
-		TimeScale,
-		CategoryScale,
-		LinearScale,
-		Title,
-		Tooltip,
-		Legend
-	);
 
 	const { data, form }: { data: PageServerData; form: ActionData } = $props();
 	let formProcessing = $state(false);
-
-	// FIXME : do proper input validation : is data null ? is there enough data ?
-	const dataArray = $derived.by(() =>
-		Array.from(data.weightArrayNotPromised, (w) => {
-			return { x: dateToStringChartTimeScaleFormatted(w.date), y: w.weight };
-		})
-	);
-
-	const dataArrayAvgState = $derived.by(() => {
-		const map = new SvelteMap<string, { sum: number; count: number }>();
-
-		for (const val of data.weightArrayNotPromised) {
-			const key = dateToStringChartTimeScaleFormatted(val.date);
-
-			const entry = map.get(key) ?? { sum: 0, count: 0 };
-			entry.sum += val.weight;
-			entry.count++;
-			map.set(key, entry);
-		}
-
-		return [...map.entries()].map(([x, { sum, count }]) => ({
-			x,
-			y: sum / count
-		}));
-	});
-
-	const firstInput = $derived(() => {
-		return { x: dataArray.at(0)!.x, y: data.userInfo!.goalWeight! };
-	});
-	const lastInput = $derived(() => {
-		return { x: dataArray.at(-1)!.x, y: data.userInfo!.goalWeight! };
-	});
-	const goalWeightDataArray = $derived([firstInput(), lastInput()]);
-
-	type TimePoint = {
-		x: string; // or Date
-		y: number;
-	};
-
-	let canvas: HTMLCanvasElement;
-	let myChart: Chart<'line', TimePoint[]> | null = null;
-	onMount(() => {
-		myChart = createMyChart(dataArrayAvgState, goalWeightDataArray);
-	});
-
-	$effect(() => {
-		if (myChart) {
-			myChart.destroy();
-			myChart = createMyChart(dataArrayAvgState, goalWeightDataArray);
-			myChart.render();
-		}
-	});
-
-	function createMyChart(input: TimePoint[], goalWeightDataArray: TimePoint[]) {
-		return new Chart(canvas, {
-			type: 'line',
-			data: {
-				datasets: [
-					{
-						label: 'Weight',
-						data: input,
-						backgroundColor: 'rgba(255, 255, 255, 1)',
-						borderColor: '#36A2EB',
-						borderWidth: 1.2,
-						pointStyle: 'rectRot',
-						pointRadius: 4,
-						tension: 0.3
-					},
-					{
-						label: 'Goal weight',
-						data: goalWeightDataArray,
-						backgroundColor: 'rgba(114, 190, 0, 1)',
-						borderColor: 'rgba(114, 190, 0, 1)',
-						borderWidth: 1,
-						pointStyle: 'line'
-					}
-				]
-			},
-			options: {
-				responsive: true,
-				maintainAspectRatio: false,
-				scales: {
-					x: {
-						title: {
-							display: true,
-							text: 'Date',
-							color: 'rgba(255,255,255, 0.6)'
-						},
-						type: 'time',
-						time: {
-							unit: 'day'
-						},
-						grid: { color: 'rgba(255,255,255, 0.1)' },
-						ticks: {
-							color: 'rgba(255,255,255, 0.6)',
-							font: {
-								size: 16
-							}
-						}
-					},
-					y: {
-						title: {
-							display: false,
-							text: 'Kilogrammes'
-						},
-						beginAtZero: false,
-						grid: { color: 'rgba(255,255,255, 0.1)' },
-						ticks: {
-							color: '#36A2EB',
-							font: {
-								size: 14
-							}
-						}
-					}
-				},
-				plugins: {
-					title: {
-						text: 'Weight graph over time',
-						display: false,
-						color: 'white',
-						font: {
-							size: 20
-						}
-					},
-					legend: {
-						labels: {
-							usePointStyle: true,
-							// pointStyleWidth: 32,
-							font: {
-								// size: 16
-							}
-						}
-					},
-					tooltip: {
-						usePointStyle: true,
-						callbacks: {
-							title: (tooltipItems) => {
-								const formatted = new Intl.DateTimeFormat('en-US', {
-									month: 'short',
-									day: '2-digit',
-									year: 'numeric'
-								}).format(new Date(tooltipItems[0].parsed.x!));
-								return formatted;
-							}
-						}
-					}
-				}
-			}
-		});
-	}
 </script>
 
 <h1 class="text-5xl">Weight</h1>
-<!-- h-[175.5px]  -->
-<section class="relative flex h-[35dvh] w-full justify-center">
-	<canvas bind:this={canvas} class="max-w-full"></canvas>
-</section>
+<WeightChart weightArray={data.weightArrayNotPromised} goalWeight={data.userInfo!.goalWeight!}
+></WeightChart>
 
 <section id="addWeight">
 	<h2 class="text-2xl">Add new log entry</h2>
